Refactor City and Gym classes; add Country entity
Updated the `City` class to include `Country` and `CountryId` properties, along with a collection of `Trainers`.
The `Gym` class was significantly refactored to replace several string properties with private setters and to change the `City` property to be of type `City`.
A new `DbSet<Country>` was added to `FitCircleDbContext` for managing `Country` entities, and a new `Country` class was introduced with properties for `Name`, `CountryCode`, and a collection of `Cities`.